:art: Type관련 배치 수정 type폴더내 ts파일에서 관리
type관련 배치 수정 type폴더내 ts파일에서 관리
