+++ houssg/src/assets/styles/StyledComponents.ts
@@ -1,8 +1,8 @@
 import styled from 'styled-components';
 
 import { color } from '.';
-import { HeightProps } from '../../types/common';
 
+//Auth
 export const AuthTitle = styled.div`
 	width: 100%;
 	padding: 0.5rem;
@@ -20,6 +20,30 @@ export const AuthContainer = styled.div`
 	align-items: center;
 `;
 
-export const HeightPositioningDiv = styled.div<HeightProps>`
-	height: ${(props) => props.height};
+//Manage관련
+
+export const HouseInfoContainer = styled.div`
+	display: flex;
+	flex-direction: row;
+	justify-content: space-between;
+	align-items: stretch;
+`;
+
+export const InfoWrapper = styled.div`
+	/* margin-top: 0.6rem; */
+	display: flex;
+	flex-direction: column;
+	text-align: left;
+	gap: 0.3rem;
+	padding: 0.5rem;
+`;
+
+export const SubInfoAligner = styled.div`
+	display: flex;
+	flex-direction: row;
+	gap: 5vw;
+`;
+
+export const InfoText = styled.div`
+	font-size: 1rem;
 `;
