+++ houssg/src/components/auth/Login.tsx
@@ -2,13 +2,13 @@ import { useNavigate } from 'react-router-dom';
 import { useState } from 'react';
 import styled from 'styled-components';
 
-import { AuthProps } from '../../types/auth';
 import { AuthContainer, AuthTitle } from '../../assets/styles';
 import { kakaoLogin } from '../../assets/images';
 import { AuthInput, AuthModeBtn, AuthSubmitBtn } from './element';
 import { usePathname } from '../../hooks';
 import { authLoginFunc } from '../../utils';
 import { KakaoAuthUri } from '../../api';
+import { AuthProps } from '../../types';
 
 const Login: React.FC<AuthProps> = ({ isLoginComp, setIsLoginComp }) => {
 	const pathname = usePathname();
