--- houssg/src/types/common/index.ts
@@ -1,5 +1,0 @@
-interface HeightProps {
-	height: string;
-}
-
-export type { HeightProps };
