+++ houssg/src/types/index.ts
@@ -1,1 +1,6 @@
 // type 지정을 위함 type 혹은 interface는 이 폴더에서 관리합니다
+
+import { ManageHouseProps } from './manage';
+import { AuthProps, AuthInputType, AuthModeType } from './auth';
+
+export type { ManageHouseProps, AuthProps, AuthInputType, AuthModeType };
