+++ houssg/src/types/manage.ts
@@ -1,0 +1,5 @@
+interface ManageHouseProps {
+	setIsEditMode: React.Dispatch<React.SetStateAction<boolean>>;
+}
+
+export type { ManageHouseProps };
